👌 IMPROVE: passing tests for v1 and new v2 tests
