+++ tests/test_cloud_run_v2.py
@@ -28,7 +28,7 @@ jobs_return_value = {
     "launchStage": "launch-stage",
     "binaryAuthorization": {},
     "template": {},
-    "observedGeneration": "observed-generation",
+    "observedGeneration": "1",
     "terminalCondition": {},
     "conditions": [],
     "executionCount": 1,
